+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 * Added the ability to drag and drop objects into the node graph
 
+### Changed
+* Fuzzy search is now exhaustive, so it can return better results
+
 ### Fixed
 * Fix dynamic node creation not working for overloaded methods
 * Fix entry point caching error when trying to get the value of an unset field